set up chart and chore express routers
