+++ memberRouter.js
@@ -1,0 +1,3 @@
+const express = require('express');
+const router = express.Router();
+const { Member } = require("./models")
